third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-css-path-expected.txt

Tests DOMNode.cssPath()

  html
    head
      head > base
    body
      body > article:nth-child(1)
      body > article:nth-child(2)
      body > input[type="number"]
      #ids
        #ids > div:nth-child(1)
        #inner-id
        #__proto__
        #\23 \22 ridiculous\22 \2e id
        #\27 quoted\2e value\27
        #\2e foo\2e bar
        #\2d
        #-a
        #\2d 0
        #\37
        #ид
        #input-id
        #ids > input[type="text"]:nth-child(13)
        #ids > p
      #classes
        #classes > div.foo.bar
        #classes > div:nth-child(2)
        #classes > div.\2e foo
        #classes > div.\2e foo\2e bar
        #classes > div.\2d
        #classes > div.-a
        #classes > div.\2d 0
        #classes > div.--a
        #classes > div.\2d --a
        #classes > div.\37
        #classes > div.класс
        #classes > div:nth-child(12)
        #classes > div.__proto__.foo
        #classes > span
        #id-with-class
        #classes > input.input-class-one
        #classes > input.input-class-two
      #non-unique-classes
        #non-unique-classes > span:nth-child(1)
        #non-unique-classes > span:nth-child(2)
        #non-unique-classes > span:nth-child(3)
        #non-unique-classes > span.c1.c2.c3
        #non-unique-classes > span:nth-child(5)
        #non-unique-classes > div:nth-child(6)
        #non-unique-classes > div.c1.c2
        #non-unique-classes > div.c3.c2
        #non-unique-classes > div.c3.c4
        #non-unique-classes > div.c1.c4
        #non-unique-classes > input:nth-child(11)
        #non-unique-classes > input:nth-child(12)
        #non-unique-classes > div:nth-child(13)
